Update README.md
Examining break

Run the program and examine the code to see how this switch statement works. The char type is used for variables that hold a single character. When using character literals in a program (like 'A') the character must be enclosed in single quotes, not double quotes that are used with string literals.

After you understand how the program functions, remove the break statement from each of the cases.

What is the effect on the execution of the program?

Include your answer as a comment at the top or bottom of the code